+++ email-log.php
@@ -5,7 +5,7 @@
  * Description: Logs every email sent through WordPress
  * Donate Link: http://sudarmuthu.com/if-you-wanna-thank-me
  * Author: Sudar
- * Version: 2.2.4
+ * Version: 2.2.5
  * Author URI: http://sudarmuthu.com/
  * Text Domain: email-log
  * Domain Path: languages/